requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
Pillow==10.4.0
pytesseract==0.3.10
sumy==0.11.0
nltk==3.9.2
